+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.4.2]
+### Improvements
+- showing a warning dialog when the application starts and **snapd** is unavailable
 ### Fixes:
 - [Snaps read index error](https://github.com/vinifmor/fpakman/issues/30)